\name{H3K36me3.AM.immune.19}
\alias{H3K36me3.AM.immune.19}
\docType{data}
\title{
 Several ChIP-seq profiles, some of which have few data points
}
\description{
  These data are used to test the PeakSegDP algorithm,
  to make sure it gives sensible results, even when there are few data.
}
\usage{data("H3K36me3.AM.immune.19")}
\format{
Named list of 21 data.frames, each with columns chromStart, chromEnd, count.
}
\source{
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/
  data set H3K36me3_AM_immune, chunk id 19
}
\keyword{datasets}
