\name{H3K36me3.TDH.other.chunk3.cluster4}
\alias{H3K36me3.TDH.other.chunk3.cluster4}
\docType{data}
\title{
8 profiles of H3K36me3 data
}
\description{
these data caused a bug in multiSampleSegHeuristic.
}
\usage{data("H3K36me3.TDH.other.chunk3.cluster4")}
\format{
  A data frame with 36914 observations on the following 4 variables.
  \describe{
    \item{\code{sample.id}}{a factor with 8 levels}
    \item{\code{chromStart}}{integer vector}
    \item{\code{chromEnd}}{integer vector}
    \item{\code{count}}{integer vector}
  }
}
\source{
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/
  data set H3K36me3_TDH_other
  chunk 3.
}
\keyword{datasets}
