\name{multiSampleSegOptimal}
\alias{multiSampleSegOptimal}
\title{multiSampleSegOptimal}
\description{Find one peak common to several samples.}
\usage{multiSampleSegOptimal(profiles)}
\arguments{
  \item{profiles}{List of data.frames with columns chromStart, chromEnd, count, or
single data.frame with additional column sample.id.}
}



\author{Toby Dylan Hocking, Guillem Rigaill}





