\name{H3K4me3.TDH.immune.chunk12.cluster4}
\alias{H3K4me3.TDH.immune.chunk12.cluster4}
\docType{data}
\title{
Histone ChIP-seq data, 26 samples, chr1 subset
}
\description{
26 samples, each with the same overlapping peak(s).
}
\usage{data("H3K4me3.TDH.immune.chunk12.cluster4")}
\format{
  A data frame.
}
\source{
http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db H3K4me3_TDH_immune data
set, chunk.id=12.
}
\keyword{datasets}
